# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "summary": "Get cart",
        "description": "Get current user's cart with all items",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_clearCart",
        "summary": "Clear cart",
        "description": "Remove all items from the cart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cart cleared successfully"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "summary": "Add item to cart",
        "description": "Add a product or variation to the cart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid item, insufficient stock, etc.)"
          },
          "404": {
            "description": "Product or variation not found"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/cart/items/{id}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "summary": "Update cart item",
        "description": "Update the quantity of an item in the cart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Cart item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (insufficient stock, etc.)"
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "summary": "Remove item from cart",
        "description": "Remove an item from the cart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Cart item ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Cart item not found"
          }
        },
        "tags": [
          "cart"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "summary": "Create order",
        "description": "Create a new order (clients can create orders)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid items, insufficient stock, etc.)"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "summary": "Get all orders",
        "description": "Get all orders (clients see only their own, admins see all)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/orders/from-cart": {
      "post": {
        "operationId": "OrdersController_createFromCart",
        "summary": "Create order from cart",
        "description": "Create a new order from the current user's cart. Cart will be cleared after order creation.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Order created successfully from cart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (cart is empty, insufficient stock, etc.)"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "summary": "Get order by ID",
        "description": "Get a specific order (clients can only access their own orders)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (client trying to access another user's order)"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "summary": "Update order",
        "description": "Update order (clients can only cancel their own orders, admins can update any order)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "OrdersController_remove",
        "summary": "Delete order",
        "description": "Delete an order (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully"
          },
          "403": {
            "description": "Forbidden (only admins can delete orders)"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "E-commerce API",
    "description": "E-commerce API with NestJS - Products, Categories, Users, and Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CartItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Cart item ID"
          },
          "productId": {
            "type": "number",
            "example": 1,
            "description": "Product ID"
          },
          "productVariationId": {
            "type": "number",
            "example": 1,
            "description": "Product Variation ID"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity in cart"
          },
          "price": {
            "type": "number",
            "example": 120,
            "description": "Current price of the product/variation"
          },
          "subtotal": {
            "type": "number",
            "example": 240,
            "description": "Subtotal (price * quantity)"
          },
          "productName": {
            "type": "string",
            "example": "Paco Rabanne",
            "description": "Product name"
          },
          "variationName": {
            "type": "string",
            "example": "XL - Green",
            "description": "Variation name if applicable"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "quantity",
          "price",
          "subtotal",
          "productName",
          "createdAt",
          "updatedAt"
        ]
      },
      "CartResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Cart ID"
          },
          "userId": {
            "type": "number",
            "example": 1,
            "description": "User ID who owns the cart"
          },
          "items": {
            "description": "Items in the cart",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 360,
            "description": "Total cart value"
          },
          "itemCount": {
            "type": "number",
            "example": 3,
            "description": "Total number of items in cart"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "userId",
          "items",
          "total",
          "itemCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "example": 1,
            "description": "Product ID (if adding base product to cart)"
          },
          "productVariationId": {
            "type": "number",
            "example": 1,
            "description": "Product Variation ID (if adding a specific variation to cart)"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity to add to cart",
            "minimum": 1
          }
        },
        "required": [
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "New quantity for the cart item",
            "minimum": 1
          }
        },
        "required": [
          "quantity"
        ]
      },
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "example": 1,
            "description": "Product ID (if ordering base product)"
          },
          "productVariationId": {
            "type": "number",
            "example": 1,
            "description": "Product Variation ID (if ordering a specific variation)"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity to order",
            "minimum": 1
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of items to order",
            "example": [
              {
                "productId": 1,
                "quantity": 2
              },
              {
                "productVariationId": 1,
                "quantity": 1
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "paymentMethod": {
            "type": "string",
            "example": "Credit Card",
            "description": "Payment method"
          },
          "shippingAddress": {
            "type": "string",
            "example": "123 Main St",
            "description": "Shipping address"
          },
          "shippingCity": {
            "type": "string",
            "example": "New York",
            "description": "Shipping city"
          },
          "shippingPostalCode": {
            "type": "string",
            "example": "10001",
            "description": "Shipping postal code"
          },
          "shippingCountry": {
            "type": "string",
            "example": "USA",
            "description": "Shipping country"
          },
          "notes": {
            "type": "string",
            "example": "Please handle with care",
            "description": "Order notes"
          }
        },
        "required": [
          "items"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Order item ID"
          },
          "productId": {
            "type": "number",
            "example": 1,
            "description": "Product ID"
          },
          "productVariationId": {
            "type": "number",
            "example": 1,
            "description": "Product Variation ID"
          },
          "price": {
            "type": "number",
            "example": 120,
            "description": "Price at time of order"
          },
          "quantity": {
            "type": "number",
            "example": 2,
            "description": "Quantity ordered"
          },
          "subtotal": {
            "type": "number",
            "example": 240,
            "description": "Subtotal for this item"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "productId",
          "price",
          "quantity",
          "subtotal",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Order ID"
          },
          "orderNumber": {
            "type": "string",
            "example": "ORD-2024-001234",
            "description": "Unique order number"
          },
          "userId": {
            "type": "number",
            "example": 1,
            "description": "User ID who placed the order"
          },
          "items": {
            "description": "Order items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "status": {
            "type": "string",
            "example": "order_created",
            "description": "Order status",
            "enum": [
              "order_created",
              "order_received",
              "order_accepted",
              "order_rejected",
              "order_shipped",
              "order_delivered",
              "order_delayed",
              "order_cancelled"
            ]
          },
          "total": {
            "type": "number",
            "example": 360,
            "description": "Total order amount"
          },
          "paymentMethod": {
            "type": "string",
            "example": "Credit Card",
            "description": "Payment method"
          },
          "paymentStatus": {
            "type": "string",
            "example": "paid",
            "description": "Payment status"
          },
          "paymentReference": {
            "type": "string",
            "example": "PAY-123456",
            "description": "Payment reference"
          },
          "shippingAddress": {
            "type": "string",
            "example": "123 Main St",
            "description": "Shipping address"
          },
          "shippingCity": {
            "type": "string",
            "example": "New York",
            "description": "Shipping city"
          },
          "shippingPostalCode": {
            "type": "string",
            "example": "10001",
            "description": "Shipping postal code"
          },
          "shippingCountry": {
            "type": "string",
            "example": "USA",
            "description": "Shipping country"
          },
          "notes": {
            "type": "string",
            "example": "Please handle with care",
            "description": "Order notes"
          },
          "receivedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order received date"
          },
          "acceptedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order accepted date"
          },
          "rejectedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order rejected date"
          },
          "shippedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order shipped date"
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order delivered date"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Order cancelled date"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "userId",
          "items",
          "status",
          "total",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "order_accepted",
            "description": "Order status",
            "enum": [
              "order_created",
              "order_received",
              "order_accepted",
              "order_rejected",
              "order_shipped",
              "order_delivered",
              "order_delayed",
              "order_cancelled"
            ]
          },
          "paymentMethod": {
            "type": "string",
            "example": "Credit Card",
            "description": "Payment method"
          },
          "paymentStatus": {
            "type": "string",
            "example": "paid",
            "description": "Payment status"
          },
          "paymentReference": {
            "type": "string",
            "example": "PAY-123456",
            "description": "Payment reference"
          },
          "notes": {
            "type": "string",
            "example": "Please handle with care",
            "description": "Order notes"
          }
        }
      }
    }
  }
}